+++ src/Pages/Dashboard/Components/index.js
@@ -3,3 +3,5 @@ export { default as TopValidators } from './TopValidators';
 export { default as TxHistory } from './TxHistory';
 export { default as RecentBlocks } from './RecentBlocks';
 export { default as RecentTxs } from './RecentTxs';
+export { default as PriceHistory } from './PriceHistory';
+export { default as HashDashboard } from './HashDashboard';